refactor(mesh): rename gossip files, consolidate transport limits
Three small structural cleanups to match the v2 design layout:

1. Rename `controller.rs` → `gossip_controller.rs` and its struct
   `MeshController` → `GossipController`. Rename `ping_server.rs` →
   `gossip_service.rs`. The old names predated the v2 design and
   actively misled: `controller.rs` was specifically the *outbound*
   gossip loop, and `ping_server.rs` is the full inbound gRPC
   `Gossip` service (sync_stream + ping + ping_req), not a ping
   handler. New names make directionality explicit.

2. New `transport/limits.rs` consolidates the wire-shape constants
   previously spread across `flow_control.rs` and `chunking.rs`:
   - `MAX_MESSAGE_SIZE` (10 MiB tonic encode/decode cap)
   - `STREAM_CHUNK_OVERHEAD_MARGIN` (64 KiB envelope headroom)
   - `MAX_STREAM_CHUNK_BYTES` (derived)
   - `DEFAULT_MAX_CHUNKS_PER_BATCH`
   - `STREAM_IDLE_TIMEOUT` (60s) — was defined inline *twice*,
     once each in the inbound and outbound sync_stream loops.

3. Delete `flow_control.rs`. The file name was misleading (neither
   piece did flow control in the networking sense) and it mashed
   two unrelated concerns together. `MAX_MESSAGE_SIZE` moved into
   `transport::limits`; `RetryManager` + `ExponentialBackoff` are
   inlined as private types at the bottom of `gossip_controller.rs`
   — their only caller. The shared `Arc<RwLock<...>>` wrapping is
   dropped since the per-peer state is owned by a HashMap and
   accessed by `&` reference; no other thread can reach it.

Build: `cargo clippy -p smg-mesh --all-targets` clean.
Tests: 112 smg-mesh lib tests pass.
