+++ crates/mesh/src/service.rs
@@ -15,7 +15,7 @@ use tonic::{
 };
 use tracing as log;
 
-use crate::flow_control::MAX_MESSAGE_SIZE;
+use crate::transport::limits::MAX_MESSAGE_SIZE;
 
 pub mod gossip {
     #![allow(unused_qualifications, clippy::absolute_paths)]
@@ -28,10 +28,10 @@ use gossip::{
 };
 
 use crate::{
-    controller::MeshController,
+    gossip_controller::GossipController,
+    gossip_service::GossipService,
     mtls::{MTLSConfig, MTLSManager},
     partition::PartitionDetector,
-    ping_server::GossipService,
 };
 
 pub type ClusterState = Arc<RwLock<BTreeMap<String, NodeState>>>;
@@ -266,8 +266,8 @@ impl MeshServer {
         .with_mesh_kv(self.mesh_kv.clone())
     }
 
-    fn build_controller(&self) -> MeshController {
-        MeshController::new(
+    fn build_controller(&self) -> GossipController {
+        GossipController::new(
             self.state.clone(),
             self.advertise_addr,
             &self.self_name,
