+++ crates/mesh/src/tests/chunking_integration.rs
@@ -34,12 +34,10 @@ use bytes::Bytes;
 use tokio::sync::mpsc::error::TryRecvError;
 
 use crate::{
-    chunking::{
-        build_stream_batches, chunk_value, dispatch_stream_batch, next_generation,
-        DEFAULT_MAX_CHUNKS_PER_BATCH, MAX_STREAM_CHUNK_BYTES,
-    },
+    chunking::{build_stream_batches, chunk_value, dispatch_stream_batch, next_generation},
     kv::{MeshKV, StreamConfig, StreamRouting},
     service::gossip::StreamEntry,
+    transport::limits::{DEFAULT_MAX_CHUNKS_PER_BATCH, MAX_STREAM_CHUNK_BYTES},
 };
 
 /// Test-scoped chunk bytes. Smaller than production to keep test
